requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0.3
djangorestframework==3.14.0
Faker==12.0.1
gunicorn==21.2.0
iniconfig==2.0.0
mixer==7.2.2
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
